pyproject.toml
[project]
name = "aimodels"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "asyncpg>=0.30.0",
    "asyncssh>=2.21.0",
    "boto3==1.37.35",
    "chromadb>=1.0.16",
    "cohere>=5.18.0",
    "fastapi>=0.116.1",
    "google-generativeai>=0.8.0",
    "groq>=0.31.1",
    "mistralai==1.7.0",
    "openai>=2.0.1",
    "opencv-python>=4.12.0.88",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "pdf2image>=1.17.0",
    "pdfplumber==0.11.6",
    "pillow==11.2.1",
    "pymupdf==1.25.5",
    "pypdf>=6.0.0",
    "pypdf2>=3.0.1",
    "pytesseract==0.3.13",
    "python-dotenv==1.1.0",
    "rapidfuzz==3.13.0",
    "rapidocr>=3.4.0",
    "rapidocr-onnxruntime>=1.4.4",
    "requests>=2.32.0",
    "sentence-transformers>=5.1.0",
    "sshtunnel>=0.4.0",
    "uvicorn>=0.35.0",
]
